[Fix] KImage width, height attribute must be integer.

+++ lib/models/base/image.dart
@@ -7,10 +7,10 @@ class KImage {
   final ImageType type;
 
   /// 이미지 객체의 가로 길이를 설정합니다.
-  final double width;
+  final int width;
 
   /// 이미지 객체의 세로 길이를 설정합니다.
-  final double height;
+  final int height;
 
   const KImage._(
     this.type, {
@@ -22,15 +22,15 @@ class KImage {
         _data = data;
 
   /// Assets으로 이미지 객체를 생성합니다.  
-  factory KImage.fromAsset(String asset, double width, double height) =>
+  factory KImage.fromAsset(String asset, int width, int height) =>
       KImage._(ImageType.assets, path: asset, width: width, height: height);
 
   /// 이미지 바이너리 값으로 이미지 객체를 생성합니다.
-  factory KImage.fromData(Uint8List data, double width, double height) =>
+  factory KImage.fromData(Uint8List data, int width, int height) =>
       KImage._(ImageType.data, data: data, width: width, height: height);
 
   /// 이미지 파일로 이미지 객체를 생성합니다.
-  factory KImage.fromFile(File file, double width, double height) =>
+  factory KImage.fromFile(File file, int width, int height) =>
       KImage._(ImageType.file, path: file.path, width: width, height: height);
 
   Map<String, dynamic> toMessageable() {
